requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
fiona==1.9.6
ftfy==6.2.0
geopandas==0.14.3
greenlet==3.0.3
idna==3.7
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pyproj==3.6.1
pyTelegramBotAPI==4.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
schedule==1.2.1
shapely==2.0.4
six==1.16.0
SQLAlchemy==2.0.29
telebot==0.0.5
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
